adjust(screens): changed Home Screen by Naver Screen
