--- src/screens/Home/index.ts
@@ -1,1 +1,0 @@
-export { default as Home } from './Home';
