+++ src/screens/Naver/index.ts
@@ -1,0 +1,1 @@
+export { default as Naver } from './NaverView';
