+++ src/screens/index.ts
@@ -1,2 +1,2 @@
-export * from './Home';
+export * from './Naver';
 export * from './Login';
